server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
scikit-learn
gunicorn
